requirements-ci.txt
# Install coderedcms from local directory.
-e .

# Requirements, in addition to coderedcms, needed for CI runner.
black==24.2.0
codespell
flake8
pytest-cov
pytest-django
sphinx
sphinx-wagtail-theme
